remove compiler warnings

+++ client/mysqltest.c
@@ -6665,7 +6665,7 @@ void init_re_comp(my_regex_t *re, const char* str)
     char erbuf[100];
     size_t len= my_regerror(err, re, erbuf, sizeof(erbuf));
     die("error %s, %d/%d `%s'\n",
-	re_eprint(err), len, (int)sizeof(erbuf), erbuf);
+	re_eprint(err), (int)len, (int)sizeof(erbuf), erbuf);
   }
 }
 
@@ -6721,7 +6721,7 @@ int match_re(my_regex_t *re, char *str)
     char erbuf[100];
     size_t len= my_regerror(err, re, erbuf, sizeof(erbuf));
     die("error %s, %d/%d `%s'\n",
-	re_eprint(err), len, (int)sizeof(erbuf), erbuf);
+	re_eprint(err), (int)len, (int)sizeof(erbuf), erbuf);
   }
   return 0;
 }
